+++ shallow_backup/constants.py
@@ -1,4 +1,4 @@
-class Constants:
+class ProjInfo:
 	PROJECT_NAME = 'shallow-backup'
 	VERSION = '1.3'
 	AUTHOR_GITHUB = 'alichtman'
@@ -6,9 +6,6 @@ class Constants:
 	DESCRIPTION = "Easily create lightweight backups of installed packages, dotfiles, and more."
 	URL = 'https://github.com/alichtman/shallow-backup'
 	AUTHOR_EMAIL = 'aaronlichtman@gmail.com'
-	CONFIG_PATH = '.shallow-backup'
-	INVALID_DIRS = [".Trash", ".npm", ".cache", ".rvm"]
-	PACKAGE_MANAGERS = ["gem", "brew-cask", "cargo", "npm", "pip", "brew", "apm"]
 	LOGO = """
               dP                dP dP                        dP                         dP
               88                88 88                        88                         88
@@ -18,3 +15,6 @@ class Constants:
      `88888P' dP    dP `88888P8 dP dP `88888P' 8888P Y8P     88Y8888' `88888P8 `88888P' dP   `YP `88888P' 88Y888P'
                                                                                                           88
                                                                                                           dP	"""
+
+
+ProjInfo = ProjInfo()
