+++ shallow_backup/printing.py
@@ -1,5 +1,5 @@
 from colorama import Fore, Style
-import constants as Constants
+from constants import ProjInfo
 import inquirer
 
 
@@ -7,9 +7,9 @@ def print_version_info(cli=True):
 	"""
 	Formats version differently for CLI and splash screen.
 	"""
-	version = "v{} by {} (@{})".format(Constants.VERSION,
-	                                   Constants.AUTHOR_FULL_NAME,
-	                                   Constants.AUTHOR_GITHUB)
+	version = "v{} by {} (@{})".format(ProjInfo.VERSION,
+	                                   ProjInfo.AUTHOR_FULL_NAME,
+	                                   ProjInfo.AUTHOR_GITHUB)
 	if not cli:
 		print(Fore.RED + Style.BRIGHT + "\t{}\n".format(version) + Style.RESET_ALL)
 	else:
@@ -20,7 +20,7 @@ def splash_screen():
 	"""
 	Display splash graphic, and then stylized version
 	"""
-	print(Fore.YELLOW + Style.BRIGHT + "\n" + Constants.LOGO + Style.RESET_ALL)
+	print(Fore.YELLOW + Style.BRIGHT + "\n" + ProjInfo.LOGO + Style.RESET_ALL)
 	print_version_info(False)
